port: darwin/arm64 M3 — time, identity, entropy, CPU count
`ClockGettime`, `Nanosleep`, `Getpid`/`Getppid`, the four uid/gid
getters, `Getgroups`, `Uname` and `Kill` are single libSystem calls
through `errno_ret`. `Getrandom` is `arc4random_buf` — Go's darwin
`readRandom` — which cannot fail or block, so every GRND_* flag is
already satisfied. `num_cpus` is Go's darwin `getCPUCount`,
`sysctl({CTL_HW, HW_NCPU})`; `SchedGetaffinity` stays an abort rather
than a faked mask. `rand::init` now runs in the Darwin boot.

Darwin's `CLOCK_MONOTONIC` counts time asleep — measured 5.4 days
ahead of `CLOCK_UPTIME_RAW` on a laptop that had slept — where Linux's
and Go's `nanotime` do not. `runtime::sysmon::NANOTIME_CLOCK` names the
clock that is Go's nanotime on each OS (`CLOCK_UPTIME_RAW` here:
`mach_absolute_time` with the timebase applied, within 166 ns of the
hand conversion), and `monotonic_ns` and `time::Now` both read it.
On Linux it is `CLOCK_MONOTONIC`, the id both already used.

Two comments claimed the Mach timebase is 1/1 on Apple Silicon. It
measures 125/3 (a 24 MHz counter); nothing read ticks as nanoseconds,
so both comments are corrected rather than any code. The generated
table's note that `CLOCK_UPTIME_RAW` lacks the unit conversion was
also wrong.

Verified against the host: Nanosleep(200ms) → Since 204–210 ms,
Unix() = `date +%s`, NumCPU = hw.ncpu = 12.

+++ src/os/mod.rs
@@ -1307,13 +1307,13 @@ pub fn Chdir<N: Into<string>>(name: N) -> error {
 fn syscallMode(i: FileMode) -> u32 {
     let mut o: u32 = i.Perm().0;
     if (i & ModeSetuid) != FileMode(0) {
-        o |= syscall::S_ISUID;
+        o |= u32::from(syscall::S_ISUID);
     }
     if (i & ModeSetgid) != FileMode(0) {
-        o |= syscall::S_ISGID;
+        o |= u32::from(syscall::S_ISGID);
     }
     if (i & ModeSticky) != FileMode(0) {
-        o |= syscall::S_ISVTX;
+        o |= u32::from(syscall::S_ISVTX);
     }
     // Go: "No mapping for Go's ModeTemporary (plan9 only)."
     return o;
